[Kernel/Memory] Most memory-related modules now support x86-64 paging.

+++ Kernel/Memory/PhysicalMemory.h
@@ -41,7 +41,7 @@ struct PACKED PhysicalRange {
 
 struct PACKED MemoryMap {
     PhysicalRange* entries;
-    u16            entry_count;
+    size_t         entry_count;
 
     PhysicalRange* begin() { return entries; }
     PhysicalRange* end() { return entries + entry_count; }
